Restructure JS to more consistent format
