+++ {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/compatibility/class-list-polyfill.js
