+++ {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/compatibility/svgforeverybody.js
