+++ {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/lazyimages/lazyimage.js
