+++ {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/main.js
@@ -1,60 +1,46 @@
 import 'babel-polyfill'
-import 'intersection-observer'
-import 'utils/class-list-polyfill'
-import 'utils/focus-ring'
-import 'utils/webp-detector'
 
-import { svg4everybody } from './utils/svgforeverybody'
+// Browser compatibility shims and helpers
+import './compatibility/class-list-polyfill'
+import './compatibility/webp-detector'
+import './compatibility/ios-device-checking'
+import { svg4everybody } from './compatibility/svgforeverybody'
 
-import LazyImage from './images'
-import { Navigation } from './site'
-import { externalLinks } from './utils'
-import { overflowTables } from './wysiwyg'
-import { bindAnimations } from './viewport-animation'
+// OSM modules
 import { bindCookieConsent } from './notices'
+import { bindAnimations } from './viewport-animation'
+import { bindExternalLinks } from './externallinks'
+import { setUpNavigation } from './navigation'
+import { setUpLazyImage } from './lazyimages'
+import { setUpFocusRing } from './focusring'
+import { setUpOverflowTables } from './wysiwyg'
+import { removePreloadClass } from './utils'
 
 document.addEventListener('DOMContentLoaded', () => {
-  const body = document.body || document.documentElement
-
-  externalLinks()
-  new Navigation()
-  overflowTables()
+  // Add handlers for cookie consent bar
   bindCookieConsent()
-
-  const lazyImage = document.querySelector('.js-LazyImage')
-  if (lazyImage) {
-    const lazyImages = document.querySelectorAll('.js-LazyImage')
-    const callback = (entries, observer) => {
-      Array.from(entries).forEach((entry, index) => {
-        if (entry.isIntersecting && !entry.target.dataset.activating) {
-          entry.target.dataset.activating = true
-          window.setTimeout(() => {
-            new LazyImage({ el: entry.target })
-            observer.unobserve(entry.target)
-          }, 150 * index)
-        }
-      })
-    }
-    /* eslint-disable compat/compat */
-    const observer = new IntersectionObserver(callback, {
-      threshold: 0,
-      rootMargin: `${window.innerHeight}px`
-    })
-    Array.from(lazyImages).forEach(image => observer.observe(image))
-  }
-
-  // If the device is iOS add a class to the body so we can do specific CSS for it
-  if (!!navigator.platform && /iPad|iPhone|iPod/.test(navigator.platform)) {
-    body.classList.add('is-iOS')
-  }
-
-  svg4everybody()
 
   // Bind the in viewport checking for animations
   bindAnimations()
 
-  // This class is used for making the animation duration on CSS animations 0, initially
-  setTimeout(() => {
-    body.classList.remove('util-Preload')
-  }, 500)
+  // Make all external links open in a new tab
+  bindExternalLinks()
+
+  // Set up the nav bar and mobile nav menu
+  setUpNavigation()
+
+  // Set up lazy loading for images
+  setUpLazyImage()
+
+  // Add utility classes for when tabbing is happening
+  setUpFocusRing()
+
+  // Handle wrapping of tables in wysiwyg editors
+  setUpOverflowTables()
+
+  // Use a sprite sheet for SVG
+  svg4everybody()
+
+  // Remove pre loading class, used for fixing some filer/animation issues on load
+  removePreloadClass()
 })
