+++ {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/navigation/index.js
@@ -1,0 +1,5 @@
+import Navigation from './navigation'
+
+export function setUpNavigation () {
+  new Navigation()
+}
