--- {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/site/index.js
@@ -1,1 +1,0 @@
-export { Navigation } from './navigation'
