+++ {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/viewport-animation/index.js
@@ -1,3 +1,4 @@
+import 'intersection-observer'
 import { isVisible } from '../utils/index'
 
 export function bindAnimations () {
