--- {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/vue/App.js
@@ -1,5 +1,0 @@
-import components from './components'
-
-export default {
-  components
-}
