--- {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/vue/components/frontend-switcher/FrontendSwitcher.css
@@ -1,46 +1,0 @@
-.fs-Outer {
-  position: fixed;
-  top: 95px;
-  left: 10px;
-  z-index: 10;
-
-  align-items: center;
-
-  display: flex;
-  width: 300px;
-  padding: 10px;
-
-  background-color: #fff;
-  border: 1px solid #dedede;
-  border-radius: 3px;
-  opacity: 0.3;
-  will-change: transform;
-
-  transition: opacity 0.18s ease;
-}
-
-.fs-Outer:hover {
-  opacity: 1;
-}
-
-.fs-Input {
-  position: relative;
-
-  align-items: center;
-  flex: 1;
-
-  display: flex;
-}
-
-.fs-Label {
-  flex-shrink: 0;
-
-  display: block;
-  margin-right: 10px;
-
-  font-size: 14px;
-}
-
-.fs-Switcher {
-  width: 100%;
-}
