--- {{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/assets/js/vue/components/index.js
@@ -1,7 +1,0 @@
-import FrontendSwitcher from './frontend-switcher/FrontendSwitcher.vue'
-import MobileNav from './mobile-nav/MobileNav.vue'
-
-export default {
-  FrontendSwitcher,
-  MobileNav
-}
